lib: Skip extern refs. to driver data structures.
When building the drivers as shared objects, we don't need to reference
the data structures with 'extern's.

Each driver needs to take care of registering the data structures
in the entry point functions.

This is precisely how linux kernel handles it.

Moreover, this allows the driver to initialize any other *internal*
state that it may need - in the entry point function call.

Signed-off-by: Kunal Gangakhedkar <kunal.gangakhedkar@gmail.com>