lib: disable fprint-list-udev-rules logic - for now.
With ENABLE_DYNAMIC_DRIVERS, fprint_get_drivers() is of no use - at least
in its current incarnation.

Guard the code with ENABLE_DYNAMIC_DRIVERS - so that it runs only
when buidling libfprint with statically linked drivers.

This can go away with alternative implementation for fprint_get_drivers()
when ENABLE_DYNAMIC_DRIVERS is on.

Signed-off-by: Kunal Gangakhedkar <kunal.gangakhedkar@gmail.com>